refactor(mutcat): relocate scalar classifier oracles to tests
